requirements.txt
numpy
pandas
scikit-learn
joblib
matplotlib
pyserial

